+++ Algorithms/Sorters/External/ISequentialStorage.cs
@@ -1,0 +1,11 @@
+namespace Algorithms.Sorters.External
+{
+    public interface ISequentialStorage<T>
+    {
+        public int Length { get; }
+
+        ISequentialStorageReader<T> GetReader();
+
+        ISequentialStorageWriter<T> GetWriter();
+    }
+}
