+++ Algorithms/Sorters/External/ISequentialStorageReader.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Algorithms.Sorters.External
+{
+    public interface ISequentialStorageReader<out T> : IDisposable
+    {
+        T Read();
+    }
+}
