+++ Algorithms/Sorters/External/ISequentialStorageWriter.cs
@@ -1,0 +1,9 @@
+using System;
+
+namespace Algorithms.Sorters.External
+{
+    public interface ISequentialStorageWriter<in T> : IDisposable
+    {
+        void Write(T value);
+    }
+}
